requirements.txt
Babel==2.7.0
certifi==2019.11.28
Django==2.2.2
djangorestframework==3.11.0
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
wincertstore==0.2
